gatewayApi/pyproject.toml
[project]
name = "gatewayapi"
version = "0.1.0"
description = "API Gateway para el Sistema de Gestión Energética EMS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic[email]>=2.11.7",
    "pymongo>=4.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.35.0",
    "motor>=3.4.0",
    "bcrypt==4.0.1",
]
